% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{conceptTableDistinctValues}
\alias{conceptTableDistinctValues}
\title{conceptTableDistinctValues}
\usage{
conceptTableDistinctValues(
  connectionDetails,
  schema,
  column,
  resultsLimit = 100
)
}
\arguments{
\item{connectionDetails}{}

\item{schema}{}

\item{column}{}

\item{resultsLimit}{}
}
\value{
character
}
\description{
query distinct values of a column in concept table
}
\author{
William J. O'Brien
}
