% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{descendants}
\alias{descendants}
\title{descendants}
\usage{
descendants(connectionDetails, conceptId)
}
\arguments{
\item{connectionDetails}{}

\item{conceptId}{}
}
\value{
tibble
}
\description{
descendants
}
\details{
Get descendants of a standard concept
}
