% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryConceptName}
\alias{queryConceptName}
\title{queryConceptName}
\usage{
queryConceptName(
  connectionDetails,
  schema,
  pattern,
  domain = NULL,
  concept = NULL,
  class = NULL,
  vocab = NULL,
  validity = NULL,
  cols = c("concept_id", "concept_name", "vocabulary_id", "standard_concept",
    "domain_id", "vocabulary_id", "concept_class_id")
)
}
\arguments{
\item{connectionDetails}{}

\item{schema}{}

\item{pattern}{}

\item{domain}{}

\item{concept}{}

\item{class}{}

\item{vocab}{}

\item{validity}{}

\item{cols}{}
}
\value{
tibble
}
\description{
Query the concept table
}
